+++ inst/extdata/sql-queries/v1-nt-distritos-clean-csv-view-es.sql
@@ -1,7 +1,7 @@
 CREATE OR REPLACE VIEW nt_csv_clean AS SELECT
     fecha,
-    CAST (distrito AS ZONES_ENUM) AS distrito,
-    CAST(numero_viajes AS N_TRIPS_ENUM) AS numero_viajes,
+    CAST (CASE distrito WHEN 'NA' THEN NULL ELSE distrito END AS ZONES_ENUM) AS distrito,
+    CAST (CASE numero_viajes WHEN 'NA' THEN NULL ELSE numero_viajes END AS N_TRIPS_ENUM) AS numero_viajes,
     personas,
     CAST(year AS INTEGER) AS ano,
     CAST(month AS INTEGER) AS mes,