+++ inst/extdata/sql-queries/v1-nt-municipios-clean-csv-view-en.sql
@@ -19,9 +19,10 @@ SELECT
     d.fecha AS date,
     CAST(CASE r.municipio_mitma
         WHEN 'externo' THEN 'external'
+        WHEN 'NA' THEN NULL
         ELSE r.municipio_mitma
         END AS ZONES_ENUM) AS id,
-    CAST(d.numero_viajes AS N_TRIPS_ENUM) AS n_trips,
+    CAST (CASE d.numero_viajes WHEN 'NA' THEN NULL ELSE d.numero_viajes END AS N_TRIPS_ENUM) AS n_trips,
     SUM(d.personas) AS n_persons,
     CAST(d.year AS INTEGER) AS year,
     CAST(d.month AS INTEGER) AS month,